Restore match edit functionality
A form has been created on the team_list page for editing matches.

This may be a performance problem (it requires a fair bit of Python
processing which may be slow on underpowered servers) so it may be
prudent to avoid creating a form for all TeamMatches on this page and
instead go back to having a separate interface (either a standalone page
or a modal) for this functionality. In that case, then this commit may
need to be reverted.