Clear seeds before updating them
We have a constraint that every seed in a TournamentDivision must be
unique. If team_a.seed = 1 and then later, team_b.seed is set to 1 and
team_a.seed is set to 2, and error will occur if team_b.seed is set
before team_a.seed. So we must clear them out before saving all of them
on an update.