Minor validation of team compositions
If a competitor is added to a team in the spreadsheet but is not on the
roster, we raise an error asking the user to fix it. Additionally, we
break the functionality out into a separate file to keep models.py from
growing too large.

Ideally, we should do more error checking, like making sure that no
competitor is on multiple teams, but seeing as alternates can be on
multiple teams, the logic for this is slightly more complicated than I
care to implement right now.

Closes #50