Fix an error in updating gdrive credentials
update_or_create was being called incorrectly - the previous version was
attempting get the row for which `key` and `value` both matched, which
would be incorrect if `value` had changed as it would attempt to put in
a second row with the same `key` and the different value.

The correct behaviour is to specify the new `value` in the `defaults`
parameter of `update_or_create` which specifies that if the filters
matching the row are found, then they should be updated to the values
specified in `defaults`.

See: https://stackoverflow.com/questions/50915911/

And: https://docs.djangoproject.com/en/2.0/ref/models/querysets/#django.db.models.query.QuerySet.update_or_create