Create Season model and migrations
This commit is basically two squashed into one. The first edit is that
the Season model is created as well as its migrations. The second is
responsible for adding the `season` column to the Tournament model. The
migrations automatically create the columns as a nullable field, then
create the seasons for the Tournament and editing the column to this
field, and finally making the column non-nullable.

The model is created in the database and the migrations are created
automatically using two migrations.