Create and migrate SchoolSeasonRegistration class
The SchoolSeasonRegistration class tracks the seasons a school is
registered for. Currently only a many-to-many relationship between
School and Season with one additional field: the division a school is
participating in for that season.

The SchoolTournamentRegistration class previously had a `school` field
which was attached to the school that was registered for it. This is
changed to a reference to the SchoolSeasonRegistration instead, as a
School must be registered for a Season in order to compete in
tournaments for that Season.

The migration copies the information that was previously in the
SchoolTournamentRegistration.school reference and populates a
SchoolSeasonRegistration row with the same information. The only
information that needed to be copied was the reference to the School.

We still have to ensure that there is a SchoolSeasonRegistration for a
School when they register for a Tournament. That is a future TODO.