Reduce database load evaluating team compositions
Suppose an object has a foreign key 'lighweight' - referencing
obj.lightweight will hit the database to retrieve the full model.
However, referencing obj.lightweight_id will not hit the database as it
is only returning the value of the foreign key.

What this means is that if we only care whether the value is non-null
(which is exactly what we want in this case), then we don't need to hit
the database to perform that operation if we use `obj.lightweight_id`
instead of `obj.lightweight`.