Update to use django channels 2.x
This commit successfully updates all remaining dependencies. Both the
front-end and back-end should now be in working order. Since it appears
that the inmemory channel layer is no longer supported, it is required
that users have their own redis installation and install channels-redis.
This has been added to requiremens.txt