Increase length of some SlugField columns
This addresses an issue that was faced at Brown 2019. A school came in
with a name length of 43 characters, and the SparringTeam ended up with
a slug that exceeded the default 50 character limit.

There are a total of 5 SlugField columns in the database. The following
where changed:

* School: the `name` column is part of the slug and could cause the slug
to exceed this length

* SparringTeam: includes the school portion of the slug in its own slug,
hence the necessity of increasing the length as well.

* Tournament: includes the location in the slug, and therefore could
exceeed length

The SlugField length for following tables were unchanged:

Season: Only uses has two year-values in the slug

SparringDivision: Only has the sex and skill_level in the slug and these
are fixed in length.